Impement forward proxy PUT
Before this commit, any node in the cluster could work as coordinator.
That's not something I want to allow at this point -> ie: Only nodes in
the preference list can be responsible for making changes to the version
of an object. That said, any node in the cluster can accept a PUT to any
key. The way this works now is that a node that receiveis a PUT that it
doesn't own simply forwards teh request to the appropriate node